% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_dateCol.R
\name{Add_dateCol}
\alias{Add_dateCol}
\title{Convert time and day of the year columns to Date format}
\usage{
Add_dateCol(df, time_col = "time", dayofyr_col = "dayofyr", date_col = "date")
}
\arguments{
\item{df}{data frame. A dataframe containing 'time' and 'dayofyr' columns.}

\item{time_col}{character. The name of the 'time' column (default is 'time').}

\item{dayofyr_col}{character. The name of the 'dayofyr' column (default is 'dayofyr').}

\item{date_col}{character. The name of the new 'Date' column to be created (default is 'date').}
}
\value{
The input data frame with an added 'Date' column.
}
\description{
This function converts the 'time' and 'dayofyr' columns of a data frame into a proper 'Date' column.
}
