% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLeapYear.R
\name{addLeapYear}
\alias{addLeapYear}
\title{Leap years}
\usage{
addLeapYear(data_in)
}
\arguments{
\item{data_in}{A data frame containing at least the columns \code{year} and \code{yday} (day of year).
The \code{year} column should be numeric, and the \code{yday} column should have values
representing the day of the year (1 to 365).}
}
\value{
A data frame with an additional row for the 366th day in leap years, and with updated
\code{date} and \code{month} columns.
}
\description{
This function adds the 366th day for leap years to a dataset. It duplicates the 365th day
for leap years, changes its day of the year (DOY) to 366, and updates the date and month.
}
\details{
This function checks for leap years (where \code{year \%\% 4 == 0}) and duplicates the data for day
365 as day 366. The \code{date} column is then recalculated using the \code{year} and \code{yday} (day of year),
and the \code{month} column is updated accordingly.
}
\examples{
# Example dataset
data_in <- data.frame(year = c(2020, 2020, 2021),
                      yday = c(365, 1, 365))

# Add leap year day
new_data <- addLeapYear(data_in)

}
