% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_sitepar_srad_elev.R
\name{adjust_sitepar_srad_elev}
\alias{adjust_sitepar_srad_elev}
\title{Adjust Sitepar File with Cloudiness and Terrain}
\usage{
adjust_sitepar_srad_elev(
  lat,
  lon,
  start_year = 1985,
  end_year = 2024,
  sitepar_path,
  output_path = NULL,
  plot = TRUE
)
}
\arguments{
\item{lat}{numeric. Latitude of site.}

\item{lon}{numeric. Longitude of site.}

\item{start_year}{numeric. Start year for cloudiness data.}

\item{end_year}{numeric. End year for cloudiness data.}

\item{sitepar_path}{string. Input path to \code{sitepar.in} file.}

\item{output_path}{string. Output path for updated file. If not given the provided sitepar_path file will be updated.}

\item{plot}{logical. If TRUE, plots cloudiness.}
}
\value{
Invisibly returns a data frame with monthly transmission values.
}
\description{
Updates \code{sitepar.in} using NASA POWER cloudiness and elevation/terrain data.
}
