% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_soilgrids_to_daycent.R
\name{convert_soilgrids_to_daycent}
\alias{convert_soilgrids_to_daycent}
\title{Convert SoilGrids data to DayCent format}
\usage{
convert_soilgrids_to_daycent(latitude, longitude, site, templatesite)
}
\arguments{
\item{latitude}{numeric. Latitude of the site in decimal degrees.}

\item{longitude}{numeric. Longitude of the site in decimal degrees.}

\item{site}{character. Name of the site. Must correspond to an existing folder name.}

\item{templatesite}{character. Soils.in file template name.
Must correspond to an existing soils.in file to be used as a template.}
}
\value{
The function exports a soils.in file for the site of interest that is DayCent-ready.
}
\description{
This function converts SoilGrids data into a DayCent-ready file format for specific locations.
SoilsGrids data is interpolated to fit the soil depth intervals for DayCent when necessary.
}
\details{
This function leverages the \code{fetchSoilGrids} function from the \code{soilDB} package to obtain data that is converted into
a DayCent-ready file format for specific locations. This function interpolates the data for soil layers using a
simple weighted approach based on the overlap between the SoilGrids and DayCent soil layer thickness.
The ksat is calculated using the Saxton and Rawls (2006) equation after converting SOC to OM.
This function pulls the Q50 values (medians) from SoilGrids for the relevant columns.
The templatesite name should match the file name in the site folder where DayCent is being run.
The site name must correpond to the folder where the exported file will be saved.
The returned soils.in file follows the template format, and thus includes 13 variables:
\itemize{
\item upper_depth_cm, from soils.in template file.
\item lower_depth_cm, from soils.in template file.
\item bdodQ50, interpolated from SoilGrids data.
\item wv0033Q50, interpolated from SoilGrids data.
\item wv1500Q50, interpolated from SoilGrids data.
\item evap_coeff, from soils.in template file.
\item fract_roots, from soils.in template file.
\item sandQ50, interpolated from SoilGrids data.
\item clayQ50, interpolated from SoilGrids data.
\item om, interpolated from SoilGrids data.
\item min_fract, from soils.in template file.
\item ksat, interpolated from SoilGrids data.
\item phh2oQ50, interpolated from SoilGrids data.
}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
latitude <- 45.7
longitude <- -118.6
site <- "test_site" ## this should be an existing folder within the sites folder for a DayCent run
templatesite <- "test_soils" ## this should be an existing soils.in to be used as a template
soils <- convert_soilgrids_to_daycent(latitude, longitude, site, templatesite) ## the file will be saved in the specified site folder.
\dontshow{\}) # examplesIf}
}
