% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DayCent_main.R
\name{daycent_main_cli}
\alias{daycent_main_cli}
\title{Command-line interface for DayCent main runner}
\usage{
daycent_main_cli(args = commandArgs(trailingOnly = TRUE))
}
\arguments{
\item{args}{character vector, usually commandArgs(trailingOnly = TRUE)}
}
\description{
Command-line interface for DayCent main runner
}
