% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DayCentRunSite.R
\name{DayCentRunSite}
\alias{DayCentRunSite}
\title{DayCent site run}
\usage{
DayCentRunSite(
  site,
  scen,
  run_eq = FALSE,
  run_base = FALSE,
  dc_exe_in = dc_exe,
  dc_path100_in = dc_path100,
  output_base = "few_outfiles.in",
  output_scen = "basic_outfiles.in",
  ...
)
}
\arguments{
\item{site}{character. Site name.}

\item{scen}{character. Scenario descriptor and should match the schedule file name for the experimental run.}

\item{run_eq}{logical. Default is FALSE. If TRUE, DayCent runs the equilibrium and the base scenario.}

\item{run_base}{logical. Default is FALSE. If TRUE, DayCent runs the base scenario.}

\item{dc_exe_in}{Local path to the DayCent executable.}

\item{dc_path100_in}{character. Local path to the 100 files folder for DayCent.}

\item{output_base}{character. Name of the .in file listing which outputs should be generated by DayCent for the base block. Default is "few_outlines.in".}

\item{output_scen}{character. Name of the .in file listing which outputs should be generated by DayCent for the experimental block. Default is "basic_outfiles.in".}
}
\value{
The function returns a character vector with the log of the DayCent run.
}
\description{
Wraps up equilibrium, base, and experimental run blocks, and moves outputs to an output folder.
}
\details{
This function uses runDayCent to perform one or more DayCent run blocks (equilibrium, base, and experimental). Results for run blocks are exported
and saved in the same site folder. This function will run one scenario at a time. If running an equilibrium block (run_eq = TRUE), a base block will also
be run.
The bin files are exported and saved in the site folder. Outputs listed in the outfiles.in file are exported and saved in an outputs folder within the site folder.
}
