% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DayCentRunSite_single_run.R
\name{DayCentRunSite_single_run}
\alias{DayCentRunSite_single_run}
\title{DayCent single site run}
\usage{
DayCentRunSite_single_run(
  site,
  scen,
  dc_exe_in = dc_exe,
  dc_path100_in = dc_path100,
  output_scen = "basic_outfiles.in",
  ...
)
}
\arguments{
\item{site}{character. Site name.}

\item{scen}{character. Scenario descriptor and should match the schedule file name for the experimental run.}

\item{dc_exe_in}{Local path to the DayCent executable.}

\item{dc_path100_in}{character. Local path to the 100 files folder for DayCent.}

\item{output_scen}{character. Name of the .in file listing which outputs should be generated by DayCent for the experimental block. Default is "basic_outfiles.in".}
}
\value{
The function returns a character vector with the log of the DayCent run.
}
\description{
This function runs DayCent for a single site.
}
\details{
This function uses \code{runDayCent} to perform a DayCent run for a single site. Results are exported
and saved in the same site folder. This function will run one scenario at a time.
The bin files are exported and saved in the site folder. Outputs listed in the outfiles.in file are exported and saved in an outputs folder within the site folder.
}
