% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDaymetData.R
\name{getDaymetData}
\alias{getDaymetData}
\title{Download Daymet weather data for DayCent run}
\usage{
getDaymetData(
  raw_data_path = NULL,
  site = site,
  lat = lat,
  lon = lon,
  start = start,
  end = end,
  ...
)
}
\arguments{
\item{raw_data_path}{character. Specifies the path where the Daymet files will be saved. Default saves files in temporary folder.}

\item{site}{character. The name or identifier for the site.}

\item{lat}{numeric. Latitude of the site in decimal degrees.}

\item{lon}{numeric. Longitude of the site in decimal degrees.}

\item{start}{numeric. Specifies the start year to download weather data.}

\item{end}{numeric. Specifies the end year to download weather data.}
}
\value{
A data frame with all dates from the start to the end year, including an additional 366th row for leap years,
and data for weather variables necessary for a DayCent run.
}
\description{
This function retrieves the Daymet weather data of a specific site in North America based on its latitude and longitude
using the \code{daymetr} package, including additional rows for day 366 in leap years.
The output includes weather variables necessary for a DayCent run: day of the month,
month, maximum and minimum temperature (in degrees Celsius), daily precipitation (cm), and shortwave radiation (W/m2).
}
\details{
The function downloads Daymet weather data based on a user-specified time interval and adds an additional row for day 366 in leap years.
More information about Daymet data can be found on: daymet.ornl.gov/overview.
The returned data frame includes the following variables in the order necessary for a DayCent run:
\itemize{
\item day, corresponds to the day of the month.
\item month, corresponds to the month of the year.
\item year, correponds to the year.
\item yday, corresponds to the day of the year.
\item tmax_C, maximum temperature, in C.
\item tmin_C, minimum temperature, in C.
\item prcp_cm_day, daily precipitation, in cm.
}
These seven variables are required in all weather files regardless of the DayCent version.
Nonetheless, the returned data frame also includes:
\itemize{
\item srad_Wm2, shortwave radiation, in W/m2.
\item VPD_kpa_day, water vapor pressure, in kPa.
}
}
\examples{
# Example usage
lat <- 40.58
lon <- -105.08
site <- "Sample Site"
weather <- getDaymetData(raw_data_path = NULL, site = site,
                          lat = lat, lon = lon, start = 2000, end = 2010)

}
