% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNASAPowerData.R
\name{getNASAPowerData}
\alias{getNASAPowerData}
\title{Download NASA Power weather data for DayCent run}
\usage{
getNASAPowerData(
  raw_data_path = NULL,
  site = site,
  lat = lat,
  lon = lon,
  start = start,
  end = end
)
}
\arguments{
\item{raw_data_path}{character. Specifies the path where the Daymet files will be saved. Default saves files in temporary folder.}

\item{site}{character. The name or identifier for the site.}

\item{lat}{numeric. Latitude of the site in decimal degrees.}

\item{lon}{numeric. Longitude of the site in decimal degrees.}

\item{start}{numeric. Specifies the start year to download weather data.}

\item{end}{numeric. Specifies the end year to download weather data.}
}
\value{
A data frame with all dates from the start to the end year and data for weather variables necessary for a DayCent run.
}
\description{
This function retrieves the NASA Power weather data of a specific site based on its latitude and longitude
using the \code{nasapower} package. The output includes weather variables necessary for a DayCent run: day of the month,
month, maximum and minimum temperature (in degrees Celsius), daily precipitation (cm), and shortwave radiation (W/m2).
}
\details{
The function downloads NASA Power weather data based on a user-specified time interval using data from the Agroclimatology Archive.
The returned data frame includes the following order necessary for a DayCent run:
\itemize{
\item day, corresponds to the day of the month.
\item month, corresponds to the month of the year.
\item year, correponds to the year.
\item yday, corresponds to the day of the year.
\item tmax_C, maximum temperature, in C.
\item tmin_C, minimum temperature, in C.
\item prcp_cm_day, daily precipitation, in cm.
}
These seven variables are required in all weather files regardless of the DayCent version.
Nonetheless, the returned data frame also includes:
\itemize{
\item srad_Wm2, shortwave radiation, in W/m2.
\item RH2M, relative humidity, in \%.
\item WS2M, wind speed at 2 meters, in mph.
}
}
\examples{
# Example usage
lat <- 40.58
lon <- -105.08
site <- "Sample Site"
weather <- getNASAPowerData(raw_data_path = NULL, site = site,
                            lat = lat, lon = lon, start = 2000, end = 2020)

}
