% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elevation.R
\name{getSiteElv}
\alias{getSiteElv}
\title{Get Elevation for a specific site}
\usage{
getSiteElv(lat, lon, site)
}
\arguments{
\item{lat}{numeric. Latitude of the site in decimal degrees.}

\item{lon}{numeric. Longitude of the site in decimal degrees.}

\item{site}{character. The name or identifier for the site.}
}
\value{
A \code{sf} object (simple feature) with columns for the site name, latitude, longitude,
and elevation in meters.
}
\description{
This function retrieves the elevation of a specific site based on its latitude and longitude
using the \code{elevatr} package. The output includes the elevation, latitude, and longitude of the site.
}
\details{
The function converts the latitude and longitude into a spatial object, queries an elevation
dataset using the \code{elevatr} package, and returns the elevation along with the latitude and
longitude. The coordinate reference system (CRS) used is WGS84 (EPSG:4326).
}
\examples{
# Example usage
lat <- 40.58
lon <- -105.08
site <- "Sample Site"
site_elevation <- getSiteElv(lat, lon, site)

}
