% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWeatherData.R
\name{getWeatherData}
\alias{getWeatherData}
\title{Daymet weather data}
\usage{
getWeatherData(
  raw_data_path = NULL,
  site = site,
  lat = lat,
  lon = lon,
  start = start,
  end = end,
  ...
)
}
\arguments{
\item{raw_data_path}{character. Specifies the path where the Daymet files will be saved. Default saves files in temporary folder.}

\item{site}{character. The name or identifier for the site.}

\item{lat}{numeric. Latitude of the site in decimal degrees.}

\item{lon}{numeric. Longitude of the site in decimal degrees.}

\item{start}{numeric. Specifies the start year to download weather data.}

\item{end}{numeric. Specifies the end year to download weather data.}
}
\value{
A data frame with all dates from the start to the end year, including an additional 366th row for leap years,
and data for weather variables necessary for a DayCent run.
}
\description{
This function retrieves the Daymet weather data of a specific site based on its latitude and longitude
using the \code{daymetr} package. The output includes weather variables necessary for a DayCent run: day of the month,
month, maximum and minimum temperature (in degrees Celsius), daily precipitation (cm), and shortwave radiation (W/m2).
}
\details{
The function downloads Daymet weather data based on a user-specified time interval,
and adds an additional row for day 366 in leap years.
}
\examples{
# Example usage
lat <- 40.58
lon <- -105.08
site <- "Sample Site"
weather <- getWeatherData(raw_data_path = NULL, site = site,
                          lat = lat, lon = lon, start = 2000, end = 2010)

}
