% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noBinFlag.R
\name{noBinFlag}
\alias{noBinFlag}
\title{Check for binary files}
\usage{
noBinFlag(run, runCheck)
}
\arguments{
\item{run}{character. Name of the DayCent block run. Matches the block specified in the schedule file name (eq, base, ...).}

\item{runCheck}{logical. Checks if a DayCent block run was required.}
}
\value{
The function returns a message based on the arguments. It will specify if a .bin file already exists or if it needs to be created.
}
\description{
This function checks if a binary file exists for the equilibrium or base DayCent block.
}
\details{
This function is used in "DayCentRunSite" function.
}
