% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DayCent_Ccycle_plotR.R
\name{plot_carbon_pools}
\alias{plot_carbon_pools}
\title{Plot DayCent carbon pools}
\usage{
plot_carbon_pools(data, scenario, start_year, end_year)
}
\arguments{
\item{data}{data frame. Data frame loaded from the dc_sip file.}

\item{scenario}{character. Scenario descriptor as a combination of site and the run descriptor (site_run).}

\item{start_year}{numeric. Start year for plots.}

\item{end_year}{numeric. End year for plots.}
}
\value{
This function returns plots for different categories of the C pool.
}
\description{
This function reads a data frame loaded from the dc_sip file and produces seven narrative plots of the carbon pools over time.
}
\details{
This function was built to use outputs from a DayCent run to plot carbon pools over a specified period of time.
The output from this function is a list of seven plots:
\itemize{
\item Plot 1: plant carbon stocks
\item Plot 2: litter carbon stocks
\item Plot 3: soil organic carbon stocks
\item Plot 4: rapid turnover of soil carbon stocks
\item Plot 5: total system carbon stocks
\item Plot 6: turnover soil carbon stocks
\item Plot 7: relative turnover of soil carbon stocks
}
}
