% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_DayCent_water.R
\name{plot_DayCent_water}
\alias{plot_DayCent_water}
\title{Visualizing Daily Water Balance Components}
\usage{
plot_DayCent_water(scen = scen, soildata_file)
}
\arguments{
\item{scen}{character. Path to watrbal.out file.}

\item{soildata_file}{character. Path to soils.in file.}
}
\value{
This function returns plots for different water balance components.
}
\description{
This package provides functions for visualizing daily water balance data,
including inflows, outflows, and storage components, using stacked and faceted plots.
}
\details{
This function was built to plot water balance components from DayCent outputs. This function requires the
following specific outfiles from DayCent:
\itemize{
\item daily water balance (watrbal.out)
\item daily volumetric soil water content by layer (vswc.out)
\item daily water filled pore space by layer (wfps.out)
}

This function also relies on the folder structure from the \code{DayCentRunSite} function.

The output from this function is a list of five plots:
\itemize{
\item Plot 1: Stacked inflows
\item Plot 2: Stacked outflows
\item Plot 3: Soil and snow storage
\item Plot 4: Faceted inflow-storage-outflow
\item Plot 5: Combined inflow and outflow
}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# scen <- "cc_nt"
  soildata_file <- "./soils.in"

test <- plot_DayCent_water(scen = scen,
                           soildata_file = soildata_file)
\dontshow{\}) # examplesIf}
}
