% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDayCent.R
\name{runDayCent}
\alias{runDayCent}
\title{Run DayCent}
\usage{
runDayCent(
  outfiles = "no_outfiles.in",
  site,
  run,
  dc_exe_in,
  dc_path100_in,
  single_site_logic = FALSE,
  ...
)
}
\arguments{
\item{outfiles}{character. Name of the .in file listing which outputs should be generated by DayCent. Default is "no_outfiles.in".}

\item{site}{character. Site name.}

\item{run}{character. Describes which DayCent block (equilibrium, base, or experimental) to run based on the schedule file name.}

\item{dc_exe_in}{Local path to the DayCent executable.}

\item{dc_path100_in}{character. Local path to the 100 files folder for DayCent.}

\item{single_site_logic}{logical. Default is FALSE. If TRUE the model runs directly from the site file. If False a logic tree determines the incoming and out going site file from the "run" variable.}

\item{...}{Additional arguments passed to downstream functions.}
}
\value{
The function returns a character vector with the log of the DayCent run. The bin and output files are exported and saved in the site folder.
}
\description{
This function builds a command line execution and runs the DayCent model.
}
\details{
This function was built to be used in the DayCent function. The rename_and_move_output_files function moves any output file created to an outputs folder.
}
