% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_exp_run.R
\name{sensitivity_exp_run}
\alias{sensitivity_exp_run}
\title{Temperature sensitivity function}
\usage{
sensitivity_exp_run(
  title,
  site,
  exp_list,
  dc_exe_in = dc_exe,
  dc_path100_in = dc_path100,
  run_base = FALSE,
  run_eq = FALSE,
  select_years = NULL,
  dc_var_list = NULL,
  dc_yr_cummsum_list = NULL,
  summary_list = NULL,
  summary_cummsum_list = NULL,
  harvest_var_list = NULL,
  ...
)
}
\arguments{
\item{title}{character. Name for output folder and files exported in this function.}

\item{site}{character. Site name.}

\item{exp_list}{character. Vector with one or more names for schedule files to run tipically representing different scenarios.}

\item{dc_exe_in}{Local path to the DayCent executable.}

\item{dc_path100_in}{character. Local path to the 100 files folder for DayCent.}

\item{run_base}{logical. Default is FALSE. If TRUE, DayCent runs the base scenario.}

\item{run_eq}{logical. Default is FALSE. If TRUE, DayCent runs the equilibrium scenario.}

\item{select_years}{numeric. Default is NULL. A vector that specifies the time interval (start and end year) for sensitivity analyses.}

\item{dc_var_list}{character. Default is NULL. A vector or list of one or more variables from the dc_sip.csv DayCent output.}

\item{dc_yr_cummsum_list}{character. Default is NULL. A vector or list of one or more variables from the dc_sip.csv DayCent output to analyze its cumulative response on a yearly timestep.}

\item{summary_list}{character. Default is NULL. A vector or list of one or more variables from the summary.out DayCent output.}

\item{summary_cummsum_list}{character. Default is NULL. A vector or list of one or more variables from the summary.out DayCent output to analyze its cumulative response on a yearly timestep.}

\item{harvest_var_list}{character. Default is NULL. A vector or list of one or more variables from the harvest.csv DayCent output.}
}
\value{
The function exports the combined dc_sip results and the combined harvest results in two separate tabular-formatted files (.csv).
The function also exports plots of SOC and any other variables listed as single JPEG files.
}
\description{
This function runs a list or series of schedule files and produces outputs from the dc_sip.csv, summary.out, or harvest.csv file.
This is a flexible function that allows the user to input lists of variables,
and produce plots of the model output (and save to output folder).
}
\details{
This function requires the DayCentRunSite function and requires a path to DayCent and DDList executables. The function first runs DayCent for each
scenario included in the schedule file list, and then collates all output files into combined dc_sip, summary, and harvest data frames. From the DayCent manual,
these outputs correspond to the following variables:
\itemize{
\item dc_sip.csv: Daily evaporation, transpiration, respiration, system C, and NPP.
\item summary.out: Daily climate, trace gas, and heterotrophic respiration.
\item harvest.csv: State of the system at time of harvest.
}

These combined files are used to plot variables of interested listed in the function. Combined files for dc_sip and harvest are exported in tabular format (.csv).
The function also returns a list with the DayCent log for each scenario run.
}
\examples{
\dontrun{
exp <- c("cc_nt_n6C", "cc_nt_n4C", "cc_nt_n2C", "cc_nt_0C",
          "cc_nt_2C", "cc_nt_4C", "cc_nt_6C") # list of schedule files
sensitivity_exp_run(title = "Wooster temperature grid search",
                     site = "wooster",
                     exp_list = exp,
                     dc_exe_in = dc_exe, dc_path100_in = dc_path100,
                     select_years = c(1962,1972),
                     dc_var_list = list("aglivc", "NPP"),
                     dc_yr_cummsum_list = c("NPP"),
                     harvest_var_list = NULL)
}

}
