% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_soils.R
\name{sensitivity_soils}
\alias{sensitivity_soils}
\title{Soil sensitivity function}
\usage{
sensitivity_soils(
  title,
  site,
  scen,
  exp_list,
  dc_exe_in = dc_exe,
  dc_path100_in = dc_path100,
  run_eq = FALSE,
  run_base = FALSE,
  select_years = NULL,
  dc_var_list = NULL,
  dc_yr_cummsum_list = NULL,
  harvest_var_list = NULL,
  ...
)
}
\arguments{
\item{title}{character. Name for output folder and files exported in this function.}

\item{site}{character. Site name.}

\item{scen}{character. Describes the scenario relative to the schedule file.}

\item{exp_list}{character. Vector with one or more names for soil files that will replce soil.in for the soil sensitivity.}

\item{dc_exe_in}{Local path to the DayCent executable.}

\item{dc_path100_in}{character. Local path to the 100 files folder for DayCent.}

\item{run_eq}{logical. Default is FALSE. If TRUE, DayCent runs the equilibrium scenario.}

\item{run_base}{logical. Default is FALSE. If TRUE, DayCent runs the base scenario.}

\item{select_years}{numeric. Default is NULL. A vector that specifies the time interval (start and end year) for sensitivity analyses.}

\item{dc_var_list}{character. Default is NULL. A vector or list of one or more variables from the dc_sip.csv DayCent output.}

\item{dc_yr_cummsum_list}{character. Default is NULL. A vector or list of one or more variables from the dc_sip.csv DayCent output to analyze its cumulative response on a yearly timestep.}

\item{harvest_var_list}{character. Default is NULL. A vector or list of one or more variables from the harvest.csv DayCent output.}
}
\value{
The function exports the combined dc_sip results and the combined harvest results in two separate tabular-formatted files (.csv).
The function also exports plots of SOC and any other variables listed as single JPEG files.
}
\description{
This function runs a schedule file across a list of soil files. This is a flexible function that allows the user to input lists of variables,
and produce plots of the model output (and save to output folder).
}
\details{
This function requires the DayCentRunSite function and requires a path to DayCent and DDList executables. The function first runs DayCent for a specific
scenario from a schedule file for different soils.in files, and then collates all output files into combined dc_sip, summary, and harvest data frames. From the DayCent manual,
these outputs correspond to the following variables:
\itemize{
\item dc_sip.csv: Daily evaporation, transpiration, respiration, system C, and NPP.
\item harvest.csv: State of the system at time of harvest.
}

These combined files are used to plot variables of interested listed in the function. Combined files for dc_sip and harvest are exported in tabular format (.csv).
The function also returns a list with the DayCent log for each scenario run.
}
\examples{
\dontrun{
exp <- c("./soils_n16_percent_clay.in", "./soils_n8_percent_clay.in", "./soils_0_percent_clay.in",
           "./soils_8_percent_clay.in", "./soils_16_percent_clay.in" ) # list of schedule files
sensitivity_soils(title = "Wooster clay vs sand grid search",
                   site = "wooster",
                   scen = "cc_nt",
                   exp_list = exp,
                   dc_exe_in = dc_exe, dc_path100_in = dc_path100,
                   select_years = c(2006,2012),
                   dc_var_list = list("aglivc", "NPP"),
                   dc_yr_cummsum_list = NULL,
                   harvest_var_list = NULL)
}

}
