% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_sch.R
\name{update_sch}
\alias{update_sch}
\title{Update Parameters in a DayCent .sch File}
\usage{
update_sch(
  in_sch = "pendleton.sch",
  out_sch = in_sch,
  param_names = c("Site file name"),
  param_values = c("pendleton_exp_all_active.100"),
  overwrite = FALSE
)
}
\arguments{
\item{in_sch}{Character. Path to the input \code{.sch} file to be read.}

\item{out_sch}{Character. Path to the output \code{.sch} file to be written.}

\item{param_names}{Character vector. The parameter names (e.g., "Site file name") to search for in the file.}

\item{param_values}{Character or numeric vector. The values to replace the matched parameters with. Must be the same length as \code{param_names}.}
}
\value{
Invisibly returns a character vector of parameters that were not found or had multiple matches.
}
\description{
This function reads a DayCent \code{.sch} schedule file and replaces one or more parameter
(e.g., "Site file name", "Starting year") with new values. This function will not write repeating parameters like schedule events.
It writes the modified file to a new output path.

The function detects whether each replacement value is numeric or character, and formats it accordingly.
}
\examples{
\dontrun{
update_sch(
  in_sch = "example.sch",
  out_sch = "example_updated.sch",
  param_names = c("Site file name", "Starting year"),
  param_values = c("site_new.100", 1985)
)
}

}
