% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DayCent_Ccycle_plotR.R
\name{plot_carbon_pools}
\alias{plot_carbon_pools}
\title{Plot DayCent carbon pools}
\usage{
plot_carbon_pools(data, scenario, start_year, end_year)
}
\arguments{
\item{data}{data frame. Data frame loaded from the dc_sip file.}

\item{scenario}{character. Scenario descriptor as a combination of site and the run descriptor (site_run).}

\item{start_year}{numeric. Start year for plots.}

\item{end_year}{numeric. End year for plots.}
}
\value{
This function returns plots for different categories of the C pool.
}
\description{
This function reads a data frame loaded from the dc_sip file and produces seven narrative plots of the carbon pools over time.
}
\details{
This function was built to be used in the DayCent function.
}
